Revert "Feature/147778 cobertura testes logistica"
